Move `storageCost*` methods to Costs
